requirements.txt
aiocron==1.8
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
black==24.10.0
blinker==1.9.0
certifi==2024.12.14
click==8.1.8
croniter==6.0.0
fastapi==0.115.11
Flask==3.1.0
frozenlist==1.5.0
future==1.0.0
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.46.1
typing_extensions==4.12.2
tzlocal==5.2
uvicorn==0.32.0
Werkzeug==3.1.3
yarl==1.18.3
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
gunicorn~=23.0.0